This is the python code of PART-3 of the project in Jupyter Notebook .
Explanation for 5% accuracy:

Codes for Part 3 of the project provided in the ipynb notebook displays the result for 1000 simulations with each simulation executing 1000 games.
This strategy is followed for all 3 boards scenarios with all 3 AI combinations.

To be more specific, each game is played 1000 * 1000 = 1,000,000 times which is a huge number of trials in respect to get the smallest probability asked i.e. with 5% accuracy of the outcome. This number of executions of the program depicts that the winning probabilities shown above are ~+95% accurate.

The above approach was based on the general concept that data sets with small sample sizes show more deviation as compared to data sets with large numbers of samples from the actual probabilities.